move form to Input comp to stop rerender on input update
