+++ src/components/chat/Input/Input.js
@@ -1,7 +1,30 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import io from 'socket.io-client';
 import './Input.css';
 
-const Input = ({ message, setMessage, sendMessage }) => (
+const socket = io(process.env.REACT_APP_REQ_URL);
+
+const Input = ({ setMessages, messages }) => {
+	const [message, setMessage] = useState('');
+
+  const sendMessage = event => {
+		event.preventDefault();
+
+		if (message) {
+			// sends message to server
+			socket.emit('message', message);
+			setMessage('');
+		}
+  };
+  
+  useEffect(() => {
+    // returned message from server
+		socket.on('retMsg', message => {
+			setMessages([...messages, message]);
+		});
+  });
+
+  return (
   <form className='form'>
     <input
       className='input'
@@ -15,6 +38,8 @@ const Input = ({ message, setMessage, sendMessage }) => (
       Send
     </button>
   </form>
-);
+  );
+};
+
 
 export default Input;
